✨ feat [#14] : 경매 낙찰자 선정 api 구현 - 경매 낙찰자 선정 관련 파일 추가 및 구현 - 커스텀 error 코드 추가 - 디렉토리 오타 수정 및 안쓰는 파일 삭제
