--- src/main/java/com/sparta/limited/auction_service/Auction_product/application/dto/request/AuctionProductCreateRequest.java
@@ -1,10 +1,0 @@
-package com.sparta.limited.auction_service.Auction_product.application.dto.request;
-
-import java.util.UUID;
-import lombok.Getter;
-
-@Getter
-public class AuctionProductCreateRequest {
-
-    private UUID productId;
-}
