--- src/main/java/com/sparta/limited/auction_service/auction/application/dto/request/AuctionUserCreateRequest.java
@@ -1,16 +1,0 @@
-package com.sparta.limited.auction_service.auction.application.dto.request;
-
-import java.math.BigDecimal;
-import java.time.LocalDateTime;
-import java.util.UUID;
-import lombok.Getter;
-
-@Getter
-public class AuctionUserCreateRequest {
-
-    private UUID productId;
-    private BigDecimal startingBid;
-    private LocalDateTime startTime;
-    private LocalDateTime endTime;
-
-}
