--- src/main/java/com/sparta/limited/auction_service/auction/application/dto/response/AuctionUserCreateResponse.java
@@ -1,14 +1,0 @@
-package com.sparta.limited.auction_service.auction.application.dto.response;
-
-import java.math.BigDecimal;
-import java.util.UUID;
-import lombok.Getter;
-
-@Getter
-public class AuctionUserCreateResponse {
-
-    private UUID id;
-    private UUID auctionId;
-    private Long userId;
-    private BigDecimal bid;
-}
