+++ src/main/java/com/sparta/limited/auction_service/auction/domain/repository/AuctionBidRepository.java
@@ -1,6 +1,7 @@
 package com.sparta.limited.auction_service.auction.domain.repository;
 
 import com.sparta.limited.auction_service.auction.domain.model.AuctionUser;
+import java.util.Optional;
 import java.util.UUID;
 
 public interface AuctionBidRepository {
@@ -8,4 +9,6 @@ public interface AuctionBidRepository {
     void save(AuctionUser auctionUser);
 
     boolean existsByAuctionIdAndUserId(UUID auctionId, Long userId);
+
+    Optional<AuctionUser> findFirstByAuctionIdOrderByBidDescCreatedAtAsc(UUID auctionId);
 }
