+++ src/main/java/com/sparta/limited/auction_service/auction/infrastructure/repository/AuctionBidRepositoryImpl.java
@@ -6,6 +6,7 @@ import com.sparta.limited.auction_service.auction.domain.repository.AuctionBidRe
 import com.sparta.limited.auction_service.auction.domain.repository.AuctionRepository;
 import com.sparta.limited.auction_service.auction.infrastructure.persistence.JpaAuctionBidRepository;
 import com.sparta.limited.auction_service.auction.infrastructure.persistence.JpaAuctionRepository;
+import java.util.Optional;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
@@ -24,6 +25,11 @@ public class AuctionBidRepositoryImpl implements AuctionBidRepository {
     @Override
     public boolean existsByAuctionIdAndUserId(UUID auctionId, Long userId) {
         return jpaAuctionBidRepository.existsByAuctionIdAndUserId(auctionId, userId);
+    }
+
+    @Override
+    public Optional<AuctionUser> findFirstByAuctionIdOrderByBidDescCreatedAtAsc(UUID auctionId) {
+        return jpaAuctionBidRepository.findFirstByAuctionIdOrderByBidDescCreatedAtAsc(auctionId);
     }
 
 }
