+++ src/main/java/com/sparta/limited/auction_service/auction_product/application/dto/request/AuctionProductCreateRequest.java
@@ -1,0 +1,10 @@
+package com.sparta.limited.auction_service.auction_product.application.dto.request;
+
+import java.util.UUID;
+import lombok.Getter;
+
+@Getter
+public class AuctionProductCreateRequest {
+
+    private UUID productId;
+}
