+++ src/main/java/com/sparta/limited/auction_service/auction_product/infrastructure/client/dto/ProductInfo.java
@@ -1,0 +1,20 @@
+package com.sparta.limited.auction_service.auction_product.infrastructure.client.dto;
+
+import java.math.BigDecimal;
+import java.util.UUID;
+
+public record ProductInfo(
+    UUID productId,
+    String title,
+    String description,
+    BigDecimal price
+
+) {
+
+    public static ProductInfo from(UUID productId, String title,
+        String description,
+        BigDecimal price) {
+        return new ProductInfo(productId, title, description, price);
+    }
+
+}
